backup
so turns out the cause of the issues from b41 was the new maxAlgorithm for id conflicts. though it should be causing less conflicts by using it it's actually causing alot more due to the fact that cwc not the auto config uses the same one. so whatever the autoconfig is it should be inverted to in order to be compatible